feat(MessageArea):update MessageArea with utils for formatmessagetime function
